Finished implementing AutoDB

+++ autodb/table/persistent_table.py
@@ -6,8 +6,8 @@ from sortedcontainers import SortedList, SortedDict
 from autodb.abc.table import Table
 from ..index import PersistentIndex
 from ..shard import ShardManager
-from ..utils.path import create_info_path, create_index_path
 from ..utils.io import empty_directory
+from ..utils.path import create_info_path, create_index_path
 from ..utils.serialization import load_object, dump_object
 
 
@@ -29,12 +29,13 @@ class PersistentTable(Table):
         """
 
         self.directory = directory
+        self._modified = False
         if not os.path.exists(directory):
             os.mkdir(directory)
         self.index_path = create_index_path(self.directory)
         self.info_path = create_info_path(self.directory)
         self.shard_manager = ShardManager(self.directory)
-        self.index_manager = PersistentIndex(self.index_path)
+        self.index_manager: PersistentIndex = PersistentIndex(self.index_path)
         if table_type is not None:
             self.table_type = table_type
             self.size = 0
@@ -54,19 +55,26 @@ class PersistentTable(Table):
 
     @classmethod
     def _from_file(cls, directory: str):
+        """Internal method to load a table from disk."""
         return cls(directory=directory)
 
     @classmethod
     def _new(cls, directory: str, table_type):
+        """Creates a new table with the given directory as the persistence location."""
         return cls(directory=directory, table_type=table_type)
 
-    def _persist(self):
-        dump_object(os.path.join(self.info_path, "table_type"), self.table_type)
-        dump_object(os.path.join(self.info_path, "size"), self.size)
-        dump_object(os.path.join(self.info_path, "unused_indexes"), self.unused_indexes)
-        self.index_manager.persist()
+    def commit(self) -> None:
+        """Commits any changes made to this table to disk."""
+        if self._modified:
+            dump_object(os.path.join(self.info_path, "table_type"), self.table_type)
+            dump_object(os.path.join(self.info_path, "size"), self.size)
+            dump_object(os.path.join(self.info_path, "unused_indexes"), self.unused_indexes)
+            self.index_manager.commit()
+            self.shard_manager.commit()
+            self._modified = False
 
     def _insert(self, item: object) -> None:
+        """Internal method that indexes and stores an object."""
         if len(self.unused_indexes) > 0:
             index = self.unused_indexes.pop()
         else:
@@ -76,9 +84,10 @@ class PersistentTable(Table):
             self.shard_manager.insert(item_dict)
         self.size += 1
         self.index_manager.index_item(item, index)
-        self._persist()
+        self._modified = True
 
     def _batch_insert(self, item_list: List[object]) -> None:
+        """Internal method that indexes and stores a list of objects."""
         first_item_type = type(item_list[0])
         item_dict = SortedDict()
         for item in item_list:
@@ -93,9 +102,10 @@ class PersistentTable(Table):
         self.shard_manager.insert(item_dict)
         for index, item in item_dict.items():
             self.index_manager.index_item(item, index)
-        self._persist()
+        self._modified = True
 
     def retrieve(self, **kwargs) -> [Generator[object, None, None]]:
+        """Retrieves items in this table based on the given argument descriptors."""
         if len(kwargs) == 0:
             raise ValueError
         indexes = self.index_manager.retrieve(**kwargs)
@@ -105,29 +115,42 @@ class PersistentTable(Table):
             return ([])
 
     def retrieve_all(self) -> [Generator[object, None, None]]:
+        """Retrieves all items in this table."""
         return self.shard_manager.retrieve_all()
 
-    def retrieve_valid_indexes(self) -> List[str]:
+    @property
+    def indexes(self) -> List[str]:
+        """Returns a list of all the indexes in this table."""
         return list(self.index_manager.index_map.keys())
 
+    @property
+    def modified(self) -> bool:
+        """Indicates whether this table has unsaved changes."""
+        return self._modified
+
     def delete(self, **kwargs):
+        """Removes items from this table based on the given descriptors."""
         if len(kwargs) == 0:
             raise ValueError
         indexes_to_delete = self.index_manager.retrieve(**kwargs)
-        self._delete_indexes(indexes_to_delete)
+        if indexes_to_delete:
+            self._modified = True
+            self._delete_indexes(indexes_to_delete)
 
     def _delete_indexes(self, indexes: Union[Set[int], List[int]]) -> None:
+        """Internal method to actually perform deletion."""
         indexes_to_delete = sorted(indexes)  # Sort indexes for shards
         self.size -= len(indexes_to_delete)  # Decrement the size of the table
         if indexes_to_delete:
+            self._modified = True
             items_to_delete = self.shard_manager.retrieve(indexes_to_delete)
             for item, index in zip(items_to_delete, indexes_to_delete):
                 self.index_manager.unindex_item(item, index)
             self.unused_indexes.update(indexes_to_delete)
             self.shard_manager.delete(indexes_to_delete)
-        self._persist()
 
     def clear(self):
+        """Removes any stored data relating to this table and clears out all items from this table."""
         empty_directory(self.directory)
         self.shard_manager = ShardManager(self.directory)
         self.index_manager = PersistentIndex(self.index_path)
